feat: organização dos componentes
