+++ src/components/AboutSection/index.tsx
@@ -1,0 +1,3 @@
+import AboutSection from "./AboutSection";
+
+export { AboutSection };
