+++ src/components/SkillsSection/skills.tsx
@@ -1,6 +1,8 @@
-export const skills: string[][] = [
+const skills: string[][] = [
     ["html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python", "html", "css", "javascript", "typescript", "react", "nextjs", "python","html", "css", "javascript", "typescript", "react", "nextjs", "python"],
     ["csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask", "csharp", "aspnetcore", "php", "nodejs", "express", "nestjs", "flask"],
     ["laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb", "laravel", "tailwind", "cssinjs", "chakraui", "mysql", "postgresql", "mongodb"],
     ["prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva", "prisma", "jest", "git", "figma", "canva"]
   ];
+
+export default skills;
